Fixed /reset and added a direct login page
I fixed the /reset exploit by making the function verify the user is logged in before resetting (should've done this originally). I also made the login page functional. It is a way to login without using google (incase it breaks, also for me to login if i need to because my google account is deleted). Username is anything on the whitelist and password is described in the function.